test(auth): cover injected session connections in integration
